# Improvements for LayerNorm