Inspired by [this](https://discourse.julialang.org/t/flux-layernorm-slower-than-pytorch/78084) discussion, this PR achieves some speedup in LayerNorm by a. passing the mean to `std` as an argument (there was a note to make this change after it had been implemented but it never really happened) and b. tweaking some mild implementation details (it removes a broadcasted `identity` that was proving to be a bottleneck for large array sizes, for example).

Before this PR (Julia 1.8.0-beta2.3, Flux master):
```julia
julia> l = LayerNorm(768);

julia> x = rand(Float32, 768, 128, 2);

julia> @benchmark $(l)($x)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  200.292 μs …  10.650 ms  ┊ GC (min … max):  0.00% … 96.39%
 Time  (median):     311.625 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   412.579 μs ± 625.478 μs  ┊ GC (mean ± σ):  23.83% ± 14.70%

  ▃█▄▁                                                          ▁
  ████▇▄▄▄▁▄▁▃▁▃▁▁▁▁▁▁▁▁▃▁▄▁▃▁▄▃▁▁▄▁▄▅▃▄▅▄▆▆▄▅▅▅▆▆▅▄▃▅▅▅▄▅▄▅▄▄▅ █
  200 μs        Histogram: log(frequency) by time       4.42 ms <

 Memory estimate: 2.25 MiB, allocs estimate: 24.
```

After this PR:
```julia
julia> @benchmark $(l)($x)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  167.875 μs …   5.326 ms  ┊ GC (min … max):  0.00% … 95.32%
 Time  (median):     243.792 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   282.340 μs ± 341.526 μs  ┊ GC (mean ± σ):  14.86% ± 11.55%

    █                                                            
  ▃▆█▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▁▂▂▂▁▂▂▂▂▂▂▂▂▁▁▁▂▂▂▂ ▂
  168 μs           Histogram: frequency by time         2.39 ms <

 Memory estimate: 1.50 MiB, allocs estimate: 16.
```